update-flakes (#283)
* feat: Update flake.lock dependencies

The changes in this commit update the dependencies in the flake.lock file. The main changes are:

- Update the `flake-parts` dependency to the latest version (rev: `506278e768c2a08bec68eb62932193e341f55c90`)
- Remove various unused dependencies related to `cachix`, `devenv`, `nix`, and `flake-compat`

These updates help keep the project's dependencies up-to-date and streamline the flake.lock file.

* feat: Add Terraform formatting to development environment

Adds Terraform formatting to the development environment by configuring the `terraform-fmt` formatter in the `flake.nix` file. This ensures that Terraform code is automatically formatted when working in the development environment.

Additionally, the shell hook message has been updated to provide a more welcoming and encouraging message for developers.

* feat: Add Terragrunt support and configure Prettier for JSON and YAML

This commit adds Terragrunt support to the Nix flake and configures Prettier to format JSON and YAML files in addition to the existing support for JavaScript, TypeScript, and Terraform files.

The key changes are:

- Add `terragrunt` to the list of required tools
- Configure `prettier` to include JSON and YAML files in the formatting
- Add a new `terragrunt-fmt` formatter to handle Terragrunt HCL files

These changes improve the development experience by providing consistent formatting across the codebase, including Terragrunt configuration files.

* feat: Add YAML linting and update dev environment configuration

- Add YAML linting using yamllint to the pre-commit hooks
- Update the dev environment configuration in flake.nix to include yamllint
- Add .gitattributes file to mark generated Go files as linguist-generated
- Update the dev environment name in flake.nix to be more descriptive
- Add support for the x86_64-windows system in the flake.nix file

* feat: Add YAML formatting and linting to the development environment

Adds the following improvements to the development environment:
- Enables YAML linting using `yamllint` for all YAML files
- Adds YAML formatting using `yamlfmt` for all YAML files
- Configures the shell environment to include the `GOROOT` environment variable
- Updates the default shell to use Bash with no profile or RC files
- Removes the unnecessary `name` field from the `devShells.default` configuration