feat: add tflinter module (#271)
* feat: Add support for custom container images and environment variables

- Add new parameters to the `New()` function to accept custom container image and environment variables
- Implement `setupContainer()` and `setupEnvironmentVariables()` functions to handle the setup of the container and environment variables
- Add a new default container image configuration in a YAML file
- Update the package description to provide more context on the purpose of the TFLint module

* feat: Update base Alpine configuration for TFLinter

The changes in this commit update the base Alpine configuration for the TFLinter project. The key changes are:

- Reorganize the `repositories` and `packages` sections for better readability.
- Update the `accounts` section to use the `tflinter` user and group.
- Add additional directories to the `paths` section, including `.terraform.d/plugin-cache`, `.terraform.d/plugins`, and `.terraform.d/providers`.
- Update the annotations with the latest version, vendor, and source information.

These changes aim to improve the overall structure and organization of the base Alpine configuration, ensuring it is optimized for the TFLinter project.

* ci: Add CI workflow for aws-tag-inspector

This commit adds a new GitHub Actions workflow for the aws-tag-inspector
module. The workflow includes the following steps:

1. Lint the aws-tag-inspector module using Dagger and GolangCI-Lint.
2. Run tests for the aws-tag-inspector module using Dagger.
3. Run the examples/go module for aws-tag-inspector using Dagger.

The workflow is triggered on push and pull request events for the main and
master branches, as well as on manual workflow dispatch. The workflow
uses the latest version of Dagger (0.14.0) and Go 1.22.

* feat: Update base Alpine image for Terragrunt

This commit updates the base Alpine image used for Terragrunt, optimizing it for security and performance. The key changes include:

- Updated Alpine version to 3.20
- Added additional packages like curl, bash, wget, unzip, git, tzdata, and openssl
- Configured the environment with Terragrunt-specific settings
- Improved the annotations with more details about the image