aws-tag-inspector-module (#285)
* feat: Add custom error handling with ModuleError

This commit introduces a custom error handling mechanism for the AwsTagInspector module. The changes include:

- Add the `err.go` file, which defines the `ModuleError` struct and related utility functions for creating, wrapping, and joining errors.
- The `ModuleError` struct provides a consistent way to handle and report errors, with a prefixed module name and emoji representation.
- The `scanner_aws_client.go` file is updated to use the new `ModuleError` for error handling.
- The `.gitignore` file is updated to ignore generated files related to the Dagger project.

The main purpose of these changes is to improve the error handling and reporting capabilities of the AwsTagInspector module, making it easier to debug and troubleshoot issues.

* feat: Update base Alpine and Wolfi APKO presets

The changes in this commit focus on updating the base Alpine and Wolfi APKO presets used for Terragrunt. The key changes are:

1. Simplify the `contents` section by removing unnecessary repository and package entries.
2. Standardize the `accounts` section across the presets, ensuring consistent group and user configurations.
3. Update the `environment` section to include common variables like `TZ`, `LANG`, and `LC_ALL`.
4. Add detailed annotations to provide more context about the presets, including the title, description, version, vendor, licenses, and source.

These changes aim to streamline the base image configurations, improve consistency, and provide more metadata for the presets.

* feat: Add S3 bucket exclusion logic

    
Add a new function `IsExcluded` to the `S3Scanner` struct that checks if a resource should be excluded from scanning based on the configured exclusion patterns. This allows users to exclude specific S3 buckets from the scanning process.

The function checks if the `config` field of the `S3Scanner` is not nil, and if the `s3` resource is enabled in the configuration. It then iterates through the `ExcludedResources` list and checks if the current resource ID matches any of the exclusion patterns. If a match is found, the function returns `true` along with the exclusion reason.

This change provides more flexibility in controlling which S3 buckets are scanned by the AWS Tag Inspector tool.